Replay the fixes done to get things working on-site.
Add in the supervisord configuration file - this starts the
sonic-pi server under xvfb and the node server.
See https://in-thread.sonic-pi.net/t/sonic-pi-for-standalone-installations/225/4
for info on 'headless'.
Supervisor conf goes in /etc,
To start at bootup, add the following to /etc/rc.local before the
'exit 0' line:

```
/usr/local/bin/supervisord
```
Node server:
Don't play the idle sample until the server is ready - because the server won't
necessarily respond soon enough, then the idleSamplePlaying flag will not get updated
and they won't play again.
Fix a bug where the samples for the change in person were not loaded.

control.rb - save the changed Sonic-Pi buffer, but also overwrite the init.rb with it's contents.
init.rb - play the amen break when we are ready, sleep and then send a message to the node-server.